refactor(core): extract maintenance/memory-role-helpers.ts — probe scoring + delta utilities
